# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artists:
    get:
      operationId: listArtists
      tags: [artists]
      summary: List artists
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: mood
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
    post:
      operationId: createArtist
      tags: [artists]
      summary: Create an artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistInput"
      responses:
        "201":
          description: Created artist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
  /artists/{id}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get an artist by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateArtist
      tags: [artists]
      summary: Update an artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistUpdate"
      responses:
        "200":
          description: Updated artist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /artists/{id}/profile:
    get:
      operationId: getArtistProfile
      tags: [artists]
      summary: Get a full artist profile with analysis, tags, links, media, and recent events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /artists/{id}/enrich:
    post:
      operationId: enrichArtist
      tags: [artists]
      summary: Trigger the enrichment pipeline for an artist (audio + lyric analysis)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrichArtistInput"
      responses:
        "202":
          description: Enrichment accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrichArtistResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /artists/{id}/media:
    post:
      operationId: addArtistMedia
      tags: [artists]
      summary: Add a media item (photo/video) to an artist gallery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaInput"
      responses:
        "201":
          description: Created media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistMedia"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /artists/{id}/media/{mediaId}:
    delete:
      operationId: deleteArtistMedia
      tags: [artists]
      summary: Delete an artist media item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /artists/{id}/collaborations:
    get:
      operationId: listArtistCollaborations
      tags: [artists]
      summary: List collaboration requests involving this artist (incoming & outgoing)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collaboration requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollaborationRequest"
  /collaborations:
    post:
      operationId: createCollaborationRequest
      tags: [artists]
      summary: Send a collaboration request to another artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaborationRequestInput"
      responses:
        "201":
          description: Created collaboration request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationRequest"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /collaborations/{id}:
    patch:
      operationId: updateCollaborationRequest
      tags: [artists]
      summary: Accept or decline a collaboration request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollaborationRequestUpdate"
      responses:
        "200":
          description: Updated collaboration request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollaborationRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Artist:
      type: object
      required: [id, artistName, genres, moodTags, themes, verified, createdAt]
      properties:
        id:
          type: string
        profileId:
          type: ["string", "null"]
        artistName:
          type: string
        bio:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        moodTags:
          type: array
          items:
            type: string
        themes:
          type: array
          items:
            type: string
        city:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        spotifyUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
    ArtistInput:
      type: object
      required: [artistName]
      properties:
        profileId:
          type: string
        artistName:
          type: string
          minLength: 1
        bio:
          type: string
        summary:
          type: string
        genres:
          type: array
          items:
            type: string
        moodTags:
          type: array
          items:
            type: string
        themes:
          type: array
          items:
            type: string
        city:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        imageUrl:
          type: string
        spotifyUrl:
          type: string
        instagramUrl:
          type: string
        youtubeUrl:
          type: string
        websiteUrl:
          type: string
    ArtistUpdate:
      type: object
      properties:
        artistName:
          type: string
          minLength: 1
        bio:
          type: string
        summary:
          type: string
        genres:
          type: array
          items:
            type: string
        moodTags:
          type: array
          items:
            type: string
        themes:
          type: array
          items:
            type: string
        city:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        imageUrl:
          type: string
        spotifyUrl:
          type: string
        instagramUrl:
          type: string
        youtubeUrl:
          type: string
        websiteUrl:
          type: string
        verified:
          type: boolean
...
    Event:
      type: object
      required: [id, name, eventDate, status, createdAt]
      properties:
        id:
          type: string
        venueId:
          type: ["string", "null"]
        artistId:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        eventDate:
          type: string
        status:
          type: string
          enum: [upcoming, live, past, cancelled]
        imageUrl:
          type: ["string", "null"]
        ticketUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ArtistAudioAnalysis:
      type: object
      required: [id, artistId, genres, moods, source, analyzedAt]
      properties:
        id:
          type: string
        artistId:
          type: string
        energy:
          type: ["number", "null"]
        danceability:
          type: ["number", "null"]
        valence:
          type: ["number", "null"]
        acousticness:
          type: ["number", "null"]
        instrumentalness:
          type: ["number", "null"]
        tempo:
          type: ["number", "null"]
        loudness:
          type: ["number", "null"]
        musicalKey:
          type: ["string", "null"]
        mode:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        moods:
          type: array
          items:
            type: string
        source:
          type: string
        analyzedAt:
          type: string
    ArtistLyricAnalysis:
      type: object
      required: [id, artistId, themes, keywords, source, analyzedAt]
      properties:
        id:
          type: string
        artistId:
          type: string
        themes:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        sentiment:
          type: ["string", "null"]
        sentimentScore:
          type: ["number", "null"]
        language:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        source:
          type: string
        analyzedAt:
          type: string
    ArtistTag:
      type: object
      required: [id, artistId, tag, type, score]
      properties:
        id:
          type: string
        artistId:
          type: string
        tag:
          type: string
        type:
          type: string
        score:
          type: number
        source:
          type: string
    ArtistLink:
      type: object
      required: [id, artistId, type, url, sortOrder]
      properties:
        id:
          type: string
        artistId:
          type: string
        type:
          type: string
        url:
          type: string
        label:
          type: ["string", "null"]
        sortOrder:
          type: integer
    ArtistMedia:
      type: object
      required: [id, artistId, type, url, sortOrder]
      properties:
        id:
          type: string
        artistId:
          type: string
        type:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
    ArtistProfile:
      type: object
      required: [artist, tags, links, media, recentEvents]
      properties:
        artist:
          $ref: "#/components/schemas/Artist"
        audioAnalysis:
          oneOf:
            - $ref: "#/components/schemas/ArtistAudioAnalysis"
            - type: "null"
        lyricAnalysis:
          oneOf:
            - $ref: "#/components/schemas/ArtistLyricAnalysis"
            - type: "null"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/ArtistTag"
        links:
          type: array
          items:
            $ref: "#/components/schemas/ArtistLink"
        media:
          type: array
          items:
            $ref: "#/components/schemas/ArtistMedia"
        recentEvents:
          type: array
          items:
            $ref: "#/components/schemas/Event"
    EnrichArtistInput:
      type: object
      properties:
        spotifyUrl:
          type: string
    EnrichArtistResult:
      type: object
      required: [artistId, status]
      properties:
        artistId:
          type: string
        status:
          type: string
          enum: [queued, dispatched, skipped]
        message:
          type: string
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    MediaInput:
      type: object
      required: [url]
      properties:
        type:
          type: string
          enum: [image, video]
        url:
          type: string
          minLength: 1
        thumbnailUrl:
          type: string
        caption:
          type: string
        sortOrder:
          type: integer
...
    CollaborationRequest:
      type: object
      required: [id, fromArtistId, toArtistId, status, createdAt]
      properties:
        id:
          type: string
        fromArtistId:
          type: string
        toArtistId:
          type: string
        fromArtistName:
          type: ["string", "null"]
        toArtistName:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
    CollaborationRequestInput:
      type: object
      required: [fromArtistId, toArtistId]
      properties:
        fromArtistId:
          type: string
        toArtistId:
          type: string
        message:
          type: string
    CollaborationRequestUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, declined]